netguardtester -> android-test
