+++ android-test/netguard/proguard-rules.pro
