+++ android-test/netguard/src/main/java/com/duckduckgo/vpn/network/impl/models/Packet.kt
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2022 DuckDuckGo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.duckduckgo.vpn.network.impl.models
+
+class Packet {
+    var time: Long = 0
+    var version = 0
+    var protocol = 0
+    var flags: String? = null
+    var saddr: String? = null
+    var sport = 0
+    var daddr: String? = null
+    var dport = 0
+    var data: String? = null
+    var uid = 0
+    var allowed = false
+
+    override fun toString(): String {
+        return "uid=$uid v$version p$protocol $daddr/$dport"
+    }
+}
