+++ android-test/settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "netguard-android"
+rootProject.name = "netguard-android-test"
 include ':netguard'
