--- android/netguardtester/src/main/cpp/CMakeLists.txt
@@ -1,28 +1,0 @@
-
-# For more information about using CMake with Android Studio, read the
-# documentation: https://d.android.com/studio/projects/add-native-code.html
-
-# Sets the minimum version of CMake required to build the native library.
-
-cmake_minimum_required(VERSION 3.18.1)
-
-# Declares and names the project.
-
-project("netguardtester")
-
-# Creates and names a library, sets it as either STATIC
-# or SHARED, and provides the relative paths to its source code.
-# You can define multiple libraries, and CMake builds them for you.
-# Gradle automatically packages shared libraries with your APK.
-
-add_library( # Sets the name of the library.
-             netguardtester
-
-             # Sets the library as a shared library.
-             SHARED
-
-             # Provides a relative path to your source file(s).
-             netguardtester.c)
-
-find_package(netguard REQUIRED CONFIG)
-target_link_libraries(netguardtester netguard::netguard)
